Add a filter to force saving payment methods in the Stripe platform account instead of the connected account.

+++ includes/class-wc-payments-customer-service.php
@@ -73,7 +73,8 @@ class WC_Payments_Customer_Service {
 		$customer_id = $this->payments_api_client->create_customer( $name, $email, $description );
 
 		if ( $user->ID > 0 ) {
-			$result = update_user_option( $user->ID, self::WCPAY_CUSTOMER_ID_OPTION, $customer_id );
+			$global = apply_filters( 'wcpay_force_network_saved_cards', false );
+			$result = update_user_option( $user->ID, self::WCPAY_CUSTOMER_ID_OPTION, $customer_id, $global );
 			if ( ! $result ) {
 				// Log the error, but continue since we have the customer ID we need.
 				Logger::error( 'Failed to store new customer ID for user ' . $user->ID );
@@ -155,15 +156,20 @@ class WC_Payments_Customer_Service {
 			return [];
 		}
 
-		$payment_methods = get_transient( self::PAYMENT_METHODS_TRANSIENT . $customer_id );
+		$cache_payment_methods = ! apply_filters( 'wcpay_force_network_saved_cards', false );
 
-		if ( $payment_methods ) {
-			return $payment_methods;
+		if ( $cache_payment_methods ) {
+			$payment_methods = get_transient( self::PAYMENT_METHODS_TRANSIENT . $customer_id );
+			if ( $payment_methods ) {
+				return $payment_methods;
+			}
 		}
 
 		try {
 			$payment_methods = $this->payments_api_client->get_payment_methods( $customer_id, $type )['data'];
-			set_transient( self::PAYMENT_METHODS_TRANSIENT . $customer_id, $payment_methods, DAY_IN_SECONDS );
+			if ( $cache_payment_methods ) {
+				set_transient( self::PAYMENT_METHODS_TRANSIENT . $customer_id, $payment_methods, DAY_IN_SECONDS );
+			}
 			return $payment_methods;
 
 		} catch ( API_Exception $e ) {
@@ -203,6 +209,9 @@ class WC_Payments_Customer_Service {
 	 * @param int $user_id WC user ID.
 	 */
 	public function clear_cached_payment_methods_for_user( $user_id ) {
+		if ( apply_filters( 'wcpay_force_network_saved_cards', false ) ) {
+			return; // No need to do anything, payment methods will never be cached in this case.
+		}
 		$customer_id = $this->get_customer_id_by_user_id( $user_id );
 		delete_transient( self::PAYMENT_METHODS_TRANSIENT . $customer_id );
 	}
